==> bpnet/analysis/1_motif_curation.Rmd <==
---
title: 'Curation of motifs from TSC model'
author: "Khayti Dalal"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
  html_document:
    depth: 3
    theme: sandstone
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  word_document:
    toc: yes
header-includes:
- \usepackage{fancyhdr}
- \usepackage{lipsum}
- \pagestyle{fancy}
- \fancyfoot[CO,CE]{1028-02098-001-001}
- \fancyfoot[LE,RO]{\thepage}
- \fancypagestyle{plain}{\pagestyle{fancy}}
editor_options:
  chunk_output_type: console
---

# Introduction

The goal of this analysis is to curate motifs found by the `TSC model` TF-MoDISco run such that there are no redundancies due palindromes, sequence similarity, overlapping peak regions etc. 

# Computational Setup

```{r, warning=F, message=F}
#Standard packages
library(rtracklayer) ; library(GenomicRanges); library(magrittr) ; library(Biostrings)
library(ggplot2) ; library(reshape2); library(plyranges); library(Rsamtools);library(BSgenome.Mmusculus.UCSC.mm10)
library(dplyr); library(data.table); library(patchwork); library(readr); library(testit);library(TxDb.Mmusculus.UCSC.mm10.knownGene)

#KNITR Options
setwd("/n/projects/kd2200/publication/bpnet/analysis/")
options(knitr.figure_dir="figures/1_motif_curation/", java.parameters = "- Xmx6g")

#Lab sources
source("/n/projects/kd2200/publication/bpnet/analysis/scripts/granges_common.r")
source("/n/projects/kd2200/publication/bpnet/analysis/scripts/metapeak_common.r")
source("/n/projects/kd2200/publication/bpnet/analysis/scripts/knitr_common.r")
source("/n/projects/kd2200/publication/bpnet/analysis/scripts/caching.r")
source("/n/projects/kd2200/publication/bpnet/analysis/scripts/metapeak_functions.R")
source("/n/projects/kd2200/publication/bpnet/analysis/scripts/multiplot.R")
source("/n/projects/kd2200/publication/bpnet/analysis/scripts/motif_functions.R")


#Pre-existing variables
modisco_dir <- '/n/projects/kd2200/publication/bpnet/modisco/profile/dataspec.yaml_default_fold_5/'
tasks <- c('cdx2','tead4','tfap2c','gata3','yap1')
threads <- 5

#modisco_dir <- '/n/projects/kd2200/publication/bpnet/modisco/profile/dataspec.yaml_default_fold_3/'

#modisco_dir <- '/n/projects/kd2200/publication/bpnet/modisco/profile/dataspec.yaml_default_fold_4/'
```

# Summary of motifs of interest

Based on the TF-MoDISco model, the following motifs are of interest to us:


```{r}
# motif set of interest.
motifs_of_interest.df <- data.frame(
  name = c('tfap2c_0','tfap2c_1','tead4_0','cdx2_0','gata3_1','gata3_4','gata3_5','tead4_double_1','ctcf','elf5','jun_fos','yap1_0','yap1_1'), 
  task = c('tfap2c','tfap2c','tead4','cdx2','gata3','gata3','gata3','tead4','cdx2','yap1','cdx2','yap1','yap1'),
  pattern = c(0,1,0,0,1,4,5,1,2,6,4,0,1)
)
motifs_of_interest.df

# motif set of interest for fold 3
# motifs_of_interest.df <- data.frame(
#   name = c('tfap2c_0','tfap2c_3','tead4_0','cdx2_0','cdx2_1','gata3_1','gata3_2','gata3_4','tead4_double_1','ctcf','elf5','jun_fos','yap1_0','yap1_3'), 
#   task = c('tfap2c','tfap2c','tead4','cdx2','cdx2','gata3','gata3','gata3','tead4','tead4','yap1','yap1','yap1','yap1'),
#   pattern = c(0,3,0,0,1,1,2,4,1,2,4,7,0,3)
# )
# motifs_of_interest.df

# motif set of interest for fold 4
# motifs_of_interest.df <- data.frame(
#   name = c('tfap2c_0','tfap2c_4','tead4_0','cdx2_0','gata3_1','gata3_2','tead4_double_1','ctcf','elf5','jun_fos','yap1_0','yap1_4'), 
#   task = c('tfap2c','tfap2c','tead4','cdx2','gata3','gata3','tead4','tead4','yap1','yap1','yap1','yap1'),
#   pattern = c(0,4,0,0,1,2,1,2,3,6,0,4)
# )
# motifs_of_interest.df
```

# Collect motif instances

First, extract each mapped set of instances from each pattern to allocate the non-curated set of motifs of interest.

```{r, eval = F}
cmds.vec<-lapply(1:nrow(motifs_of_interest.df), function(x){
  info.df<-motifs_of_interest.df[x,]
  cmd<-paste0('bpnet cwm-scan --patterns metacluster_0/pattern_', info.df$pattern, ' ',' ',
         modisco_dir, '/', info.df$task,' ',' ', '/n/projects/kd2200/analysis/BPNet/cegkttyz_tsc_bpnet/grid_search/final_cttgy_tsc_model/bpnet/analysis/fold_5/tsv/',
         info.df$name, '_instances-all-regions.tsv.gz')
  return(cmd)
}) %>% unlist
cmds.vec<-c('#!bin/bash', cmds.vec, '')


readr::write_lines(cmds.vec, file = '/n/projects/kd2200/publication/bpnet/analysis/tmp/tsc_model_cwm_scan_indiv_instances.sh', sep = '\n')
```

Run CWM-scanning to allocate.

```{bash, eval = F}
conda activate bpnet
<!-- bash /n/projects/kd2200/publication/bpnet/analysis/tmp/tsc_model_cwm_scan_indiv_instances.sh -->
```

Next import motifs.

```{r}
motifs.df<-lapply(1:nrow(motifs_of_interest.df), function(x){
    info.df<-motifs_of_interest.df[x,]
    df<-readr::read_tsv(paste0('/n/projects/kd2200/publication/bpnet/analysis/tsv/motif_curation/', info.df$name, '_instances-all-regions.tsv.gz')) %>%
        dplyr::mutate(pattern_name = info.df$name,
                      modisco_task = info.df$task) 
}) %>% rbindlist(fill = TRUE)
```

Report number of motifs

```{r}
motifs.df$pattern_name %>% table
```

# Remove redundant palindromic mappings

Motifs that are palindromic will be mapped on the positive and negative strand with almost identical ranges. Here, we will remove redundantly mapped motifs and select the unique motif maps that have the highest importance score.

```{r}
#Remove palindromic motifs for each pattern.
motifs.gr<-lapply(motifs.df$pattern_name %>% unique, function(x){
  message(x)
  gr<-remove_palindromic_motifs_from_bpnet_instances(dfi = motifs.df, x)
}) %>% GRangesList %>% unlist
```


# Measure %age of motifs across enhancer edges

```{r}
edge_motifs.vec<-motifs.gr %>% plyranges::filter(pattern_center <=100 | pattern_center>=900) %>% .$pattern_name %>% table
total_motifs.vec<-motifs.gr$pattern_name %>% table
round(edge_motifs.vec/total_motifs.vec, 2)
```

Less than 5% of all motifs are on mapped BPNet window edges.


Consolidate motifs.

```{r}
motifs_curated.gr<- motifs.gr %>% plyranges::arrange(pattern_name) 
motifs_curated.gr$motif_type<-"short"
motifs_curated.gr<-motifs_curated.gr %>% plyranges::mutate(motif_id = 1:length(.))
```

Plot a general summary of the motifs.

```{r TSC_curated_motif_freq, fig.height = 3, fig.width=10}
ggplot(motifs.gr %>% as.data.frame, aes(x = pattern_name))+
    geom_bar(aes(fill = factor(contrib_weighted_cat, levels = c('high','medium','low'))), color = 'black')+
    scale_fill_manual(values = rev(c('#83adb5', '#c7bbc9', '#5e3c58')), name = 'contrib')+
    scale_x_discrete(name = 'motifs')+
    ggtitle('TSC curated motif frequencies')+
    theme_classic()
```

## Double check motif overlaps

Ensure that there isn't an overrepresentation of unexpected overlapping motif pairs. We expect some overlaps, but checking to make sure that our curations are consistent. Here, we will filter out the longer motifs because we know that those overlaps will be highly represented.

```{r}
short_motifs.gr<-motifs_curated.gr %>% plyranges::filter(motif_type=='short')
overlapping_motifs.gr<- short_motifs.gr %>% 
  GenomicRanges::reduce(ignore.strand = T) %>% 
  plyranges::mutate(motif_island_id = 1:length(.))
overlapping_motifs.ov<-findOverlaps(short_motifs.gr, overlapping_motifs.gr, ignore.strand = T)
assert('IDs do not allow indexing shortcut.', 1:length(short_motifs.gr)==overlapping_motifs.ov@from)
short_motifs.gr$motif_island_id<-overlapping_motifs.ov@to

#Group motifs by their island ID and display frequency of motif contents
grouped_motif_content.df<-short_motifs.gr %>% as.data.frame %>% 
  dplyr::group_by(motif_island_id) %>% 
  dplyr::summarize(motif_content = paste0(sort(pattern_name), collapse = '_'))

grouped_motif_content.df$motif_content %>% table() %>% sort(decreasing = T) %>% .[1:30]
```

Here we can see that the most frequent motif overlaps occur between Bcd and the other motifs, but at very low rates which will not throw off analysis.

# Annotate motifs by their collected regional grouping

Because we took motifs based on their relative contribution and not the featured `example_idx` column, we need a new grouping annotation to define the motifs when we are conducting perturbations and other grouping systems. We will call this index `region_id`.

```{r}
islands.gr<-resize(motifs_curated.gr, 200, 'center') %>% 
  GenomicRanges::reduce(ignore.strand = T) %>% 
  plyranges::mutate(region_id = 0:(length(.)-1)) #Make region_id 0-based in order to be compatible with BPNet indexing properties.
ov<-findOverlaps(motifs_curated.gr, islands.gr, ignore.strand = T)
testit::assert(ov@from %>% unique %>% length == ov@from %>% length)

#Show hoy many motifs there are per window
ov@to %>% table %>% table

#Assign region_id to these motifs
motifs_curated.gr$region_id<-islands.gr$region_id[ov@to]
motifs_curated.gr$region_start_1based<-islands.gr[ov@to] %>% resize(., 1000, 'center') %>% start()
motifs_curated.gr$region_end_1based<-islands.gr[ov@to] %>% resize(., 1000, 'center') %>% end()

#Assert that all motifs are incorporated.
testit::assert('All motifs have a correct region', all(overlapsAny(motifs_curated.gr, resize(islands.gr, 1000, 'center'))))

#Export regions in 0-based coordinates
islands.gr<-islands.gr %>% resize(., 1000, 'center') 
rtracklayer::export(islands.gr %>% plyranges::mutate(name = region_id), 'bed/curated_island_regoion_tsc_model_motifs_fold5_0based.bed', format = 'BED') 
```

# Export curated motifs

In the subsequent .bed file it will be the `[task]_[motif_id]_[region_id]` naming convention.

```{r}

motifs_curated.df<-motifs_curated.gr %>% as.data.frame()
readr::write_tsv(motifs_curated.df, 'tsv/curated_motif_regoion_tsc_model_motifs_fold5_1based.tsv.gz')
rtracklayer::export(motifs_curated.gr %>% plyranges::mutate(name = paste0(pattern_name, '_', region_id)), 
                    'bed/curated_motif_regoion_tsc_model_motifs_fold5_0based.bed', format = 'BED')

mm10_promoters <-promoters(TxDb.Mmusculus.UCSC.mm10.knownGene,upstream=1000, downstream=200)
rm <- readRDS("/n/projects/kd2200/publication/bpnet/analysis/data/repeatmasker.mm10.gr.rds")
erv <- rm[grep("ERV", rm$repeat_class)]
gr_no_erv <- motifs_curated.gr[which(!overlapsAny(motifs_curated.gr, erv, ignore.strand=T))]
gr_no_erv_no_promo <- gr_no_erv[which(!overlapsAny(gr_no_erv, mm10_promoters, ignore.strand=T))]

##noerv_nopromo
motifs_curated2.df<-gr_no_erv_no_promo %>% as.data.frame()
readr::write_tsv(motifs_curated2.df, 'tsv/curated_motif_regoion_tsc_model_motifs_fold5_1based_noerv_nopromoter.tsv.gz')
rtracklayer::export(gr_no_erv_no_promo %>% plyranges::mutate(name = paste0(pattern_name, '_', region_id)), 
                    'bed/curated_motif_regoion_tsc_model_motifs_fold5_0based_noerv_nopromoter.bed', format = 'BED')

##noerv
motifs_curated3.df<-gr_no_erv %>% as.data.frame()
readr::write_tsv(motifs_curated3.df, 'tsv/curated_motif_regoion_tsc_model_motifs_fold5_1based_noerv.tsv.gz')
rtracklayer::export(gr_no_erv %>% plyranges::mutate(name = paste0(pattern_name, '_', region_id)), 
                    'bed/curated_motif_regoion_tsc_model_motifs_fold5_0based_noerv.bed', format = 'BED')

```

In conclusion, motifs were curated based on redundancies and saved to 1-based and 0-based coordinates. These will be the reference motifs that we base subsequent analysis on.

# Session Information

For the purposes of reproducibility, the R/Bioconductor session information is printed below:

```{r sessioninfo}
sessionInfo()
```
