==> bpnet/analysis/4a_pairwise_insilico_distance_analysis.ipynb <==


# Introduction

The purpose of this .ipynb is to generate motif-pair in-silico perturbations across canonical `TSC model` motifs and their variants. This will allow us to assess whether motif pairs have preferential distances or spacings between them.

# Computational setup





import warnings
warnings.filterwarnings("ignore")
from tensorflow.python.util import deprecation
deprecation._PRINT_DEPRECATION_WARNINGS = False

#Packages
import os
import sys
import pandas as pd
import numpy as np
from tqdm import tqdm
from pybedtools import BedTool

# Settings
os.chdir(f'/n/projects/kd2200/publication/bpnet/analysis/')
pd.set_option('display.max_columns', 100)

# Custom commands
sys.path.insert(0, f'/n/projects/kd2200/publication/bpnet/analysis/scripts/bpnet/scripts')
from data_format_functions import myround, myfloor, myceiling

# function to return key for any value 
def get_key(val, my_dict): 
    for key, value in my_dict.items(): 
        if val == value: 
            return key 
    return "key doesn't exist"

# Get tsc model directory
tsc_model_dir = f"/n/projects/kd2200/publication/bpnet/model/dataspec.yaml_default_fold_5/"


from bpnet.BPNet import BPNetSeqModel
tsc_model_bpnet = BPNetSeqModel.from_mdir(tsc_model_dir)

from collections import OrderedDict

##when Tead4 and Yap1 task are given same tead4 single motif
##tead4
# motif_seq_dict = OrderedDict([
#     ("tead4", 'GGAATGT'),    
#     ("tfap2c", 'GCCTCAGGG'),
#     ("cdx2", 'GCCATAAA'),
#     ("gata3", 'AGATAAG'),
# ])
##yap1
# motif_seq_dict = OrderedDict([
#     ("yap1", 'GGAATGT'),    
#     ("tfap2c", 'GCCTCAGGG'),
#     ("cdx2", 'GCCATAAA'),
#     ("gata3", 'AGATAAG'),
# ])

##when Tead4 and Yap1 task are given same tead4 double motif
##tead4
# motif_seq_dict = OrderedDict([
#     ("tfap2c", 'GCCTCAGGG'),
#     ("cdx2", 'GCCATAAA'),
#     ("gata3", 'AGATAAG'),
#     ("tead4", 'GGAATGCCAGGAATGT'), 
# ])

##yap1
motif_seq_dict = OrderedDict([
    ("tfap2c", 'GCCTCAGGG'),
    ("cdx2", 'GCCATAAA'),
    ("gata3", 'AGATAAG'),
    ("yap1", 'GGAATGCCAGGAATGT'), 
])

#Get each pairwise combination
#The order matters so I use permutation instead of combination
from itertools import permutations
pair_combos =[",".join(map(str, perms)) for perms in permutations(motif_seq_dict.keys(), 2)]

#Note: This does not include motif-motif combinations!
pair_combos

# Split the pairwise combination to have center motif and side motif

motif_keys = pd.DataFrame([])

for pair in pair_combos:
    keys = pair.split(",")
    motif_keys = motif_keys.append(keys)
motif_keys.head(n=5)    

# Get all pair-wise spacing simulation

from bpnet.simulate import generate_sim
spacing_pairs = pd.DataFrame([])
for pair in pair_combos:
    motif_keys = pair.split(",")
    res = generate_sim(tsc_model_bpnet, central_motif=motif_seq_dict[motif_keys[0]], 
                       side_motif=motif_seq_dict[motif_keys[1]], repeat=256,
                       side_distances=np.arange(505, 650), 
                       center_coords=[475, 525], #Window to compute signal across
                       contribution=[], correct=True)

    dfs, profiles = res
    dfs['central_motif_name'] = motif_keys[0]
    dfs['side_motif_name'] = motif_keys[1]
    spacing_pairs = spacing_pairs.append(dfs)
    spacing_pairs.head(n=10)

# Save in tsv

##when Tead4 and Yap1 task are given same tead4 single motif
#spacing_pairs.to_csv(f"tsv/4a_cttg_insilico_distance_analysis.tsv", sep="\t")
#spacing_pairs.to_csv(f"tsv/4a_ctyg_insilico_distance_analysis.tsv", sep="\t")

##when Tead4 and Yap1 task are given same tead4 double motif
#spacing_pairs.to_csv(f"tsv/4a_cttdblg_insilico_distance_analysis_tead4.tsv", sep="\t")
spacing_pairs.to_csv(f"tsv/4a_cttdblg_insilico_distance_analysis_yap1.tsv", sep="\t")


# Load the tsv

##when Tead4 and Yap1 task are given same tead4 single motif
#spacing_pairs = pd.read_csv(f"tsv/4a_cttg_insilico_distance_analysis.tsv", sep="\t")
#spacing_pairs = pd.read_csv(f"tsv/4a_ctyg_insilico_distance_analysis.tsv", sep="\t")

##when Tead4 and Yap1 task are given same tead4 double motif
#spacing_pairs = pd.read_csv(f"tsv/4a_cttdblg_insilico_distance_analysis_tead4.tsv", sep="\t")
spacing_pairs = pd.read_csv(f"tsv/4a_cttdblg_insilico_distance_analysis_yap1.tsv", sep="\t")
spacing_pairs.head(n=5)

# Create a new column for motif_pair
# The Motif1-Motif2 and Motif2-Motif1 will be grouped as same motif_pair 
# and I will only plot the center motif for these two cases
import collections
spacing_pairs['motif_pair_name'] = spacing_pairs[['central_motif_name', 'side_motif_name']].apply(lambda x: '<>'.join(sorted(x)), axis=1)
collections.Counter(spacing_pairs['motif_pair_name'])
spacing_pairs

# Get the task that match the center motif
spacing_pairs['task_and_motif_match'] = [(spacing_pairs['task'].iat[i] == spacing_pairs['central_motif_name'].iat[i]) for i in range(spacing_pairs.shape[0])]
spacing_pairs['task_and_motif_match']= spacing_pairs['task_and_motif_match'].astype(str) 
spacing_pairs_subset = spacing_pairs[spacing_pairs['task_and_motif_match'] =='True']

tf_colors = {
    "tead4": "#007d47",
    "tfap2c": "#c22f2f",
     "yap1": "#14a0b5",  
     "cdx2": "#E69F00",
     "gata3": "#3600e6",
}

from plotnine import *
import plotnine

plotnine.options.figure_size = (8, 8)
spacing_fig = (ggplot(aes(x='distance', y='profile/counts_max_ref_frac', color = 'task'), spacing_pairs_subset) +  
 geom_line(size=0.9) + 
 geom_hline(yintercept=1, alpha=0.5) + 
 scale_color_manual(values = tf_colors) + 
 scale_x_continuous(limits = [5, 150], breaks = range(0, 150, 25), name = 'Motif pair distance (bp)')+
 scale_y_continuous(limits = [0,3]) +
 facet_wrap('motif_pair_name', ncol = 2, scales = "free")+
 theme_bw())

##save plots
#spacing_fig.save('figures/4a_pairwise_insilico_distance_analysis/4a_cttg_pairwise_motif_spacing.pdf', height=10, width=10)
#spacing_fig.save('figures/4a_pairwise_insilico_distance_analysis/4a_ctyg_pairwise_motif_spacing.pdf', height=10, width=10)
#spacing_fig.save('figures/4a_pairwise_insilico_distance_analysis/4a_cttdblg_insilico_distance_analysis_tead4.pdf', height=10, width=10)
spacing_fig.save('figures/4a_pairwise_insilico_distance_analysis/4a_cttdblg_insilico_distance_analysis_yap1.pdf', height=10, width=10)
