==> bpnet/analysis/7a_tead4_cooperativity_doublemotifs_tscmodel.ipynb <==
# Introduction

This analysis to quantify the cooperative binding of Tead4 on Tead double motifs genome-wide using trined TSC model. We  with measure the fold increase in binding to the intact double motif as compared to the sum of the binding to each motif alone.

# Computational setup

## Environment

import warnings
warnings.filterwarnings("ignore")
from tensorflow.python.util import deprecation
deprecation._PRINT_DEPRECATION_WARNINGS = False

#Packages
import os
import sys
import pandas as pd
import numpy as np
import plotnine
#!pip install logomaker
import seaborn as sns
from pybedtools import BedTool
from tqdm import tqdm
from bpnet.utils import create_tf_session
from bpnet.BPNet import BPNetSeqModel
from bpnet.simulate import motif_coords
from bpnet.preproc import resize_interval
from concise.preprocessing import encodeDNA
from bpnet.extractors import extract_seq
from plotnine import *
from Bio.Seq import Seq

# Settings
os.chdir(f'/n/projects/kd2200/publication/bpnet/analysis/')
pd.set_option('display.max_columns', 100)
create_tf_session('0',1)

# Custom functions
sys.path.insert(0, f'/n/projects/kd2200/publication/bpnet/analysis/scripts/bpnet/scripts')
from perturb_functions import insert_motif, one_hot_decode_sequence, one_hot_encode_sequences, one_hot_encode_sequence
from motif_functions import remove_palindromic_motif_duplicates
from data_format_functions import tidy_bpnet_predictions_nexus,tidy_bpnet_contributions
sys.path.insert(0, f'/n/projects/kd2200/publication/bpnet/analysis/scripts/py')
from motifs import extract_seqs_from_df

## Variables

#%matplotlib inline
# Pre-existing variables
fasta_file = f'/n/projects/kd2200/publication/bpnet/fasta/mm10.fa'
model_dir = f'/n/projects/kd2200/publication/bpnet/model/dataspec.yaml_default_fold_5/'

# Independent variables
#tead4_motifs_path = f'tsv/tead4_double_motifs_no_ov_noerv_no_promoter_cwm_fold5_test.tsv.gz'
tead4_motifs_path = f'/n/projects/kd2200/publication/bpnet/cwm_all_regions/tead4/instances_by_pattern/m0_p1_motif-instances-all-regions.tsv.gz'


## Load BPNet model

model = BPNetSeqModel.from_mdir(model_dir)

## Import Tead4 double motifs

motifs_df = pd.read_csv(tead4_motifs_path, sep = '\t')
motifs_df.shape
motifs_df   

# Get repeat masker annotation and remove ERVs from dataframe

# make grnages using pyrange
import pyranges as pr

motifs_df['row_idx'] = np.arange(len(motifs_df))
motifs_df['Chromosome'] = motifs_df.example_chrom
motifs_df['Start'] = motifs_df.pattern_start_abs
motifs_df['End'] = motifs_df.pattern_end_abs
motifs_df['Strand'] = motifs_df.strand

cwm_scan_regions_pr = pr.PyRanges(motifs_df)
len(cwm_scan_regions_pr)

#functions
def read_repeat_masker(file_path):
    dfrm = pd.read_table(file_path, delim_whitespace=True, header=[1])

    dfrm.columns = [x.replace("\n", "_") for x in dfrm.columns]

    dfrm['name'] = dfrm['repeat'] + "//" + dfrm['class/family']

    dfrm = dfrm[['ins.', 'sequence', 'begin', 'name']]
    dfrm.columns = ['chrom', 'start', 'end', 'name']
    return dfrm


def intersect_repeat_masker(pattern_name, seqlets: BedTool, repeat_masker: BedTool, f=1.0):
    """Intersect the seqlets bed file with 
    """
    try:
        dfint = seqlets.intersect(repeat_masker, wa=True, wb=True, f=f).to_dataframe()
    except Exception:
        return None
    t = dfint.blockCount.str.split("//", expand=True)
    dfint['pattern_name'] = pattern_name
    dfint['repeat_name'] = t[0]
    dfint['repeat_family'] = t[1]
    dfint['n_pattern'] = seqlets.to_dataframe()[['chrom', 'start', 'end']].drop_duplicates().shape[0]
    dfint['interval'] = dfint['chrom'] + ":" + dfint['start'].astype(str) + "-" + dfint['end'].astype(str)
    return dfint[['chrom', 'start', 'end', 'interval', 'pattern_name', 'n_pattern', 'repeat_name', 'repeat_family']]

# load repeat masker file and remove ervs
rm_filepath = f'/n/projects/kd2200/publication/bpnet/analysis/data/repeatmasker.mm10.fa.out.gz'
dfrm = read_repeat_masker(rm_filepath)
dfrm_erv = dfrm[dfrm.name.str.contains("ERV")]
dfrm_erv['Start'] = dfrm_erv['start']
dfrm_erv['End'] = dfrm_erv['end']
dfrm_erv['Chromosome'] = dfrm_erv['chrom']
dfrm_erv = pr.PyRanges(dfrm_erv)

# exclude erv's
exclude = cwm_scan_regions_pr.overlap(dfrm_erv).df.row_idx.unique()
exclude_rows = motifs_df.row_idx.isin(exclude)
print(exclude_rows.value_counts())
print(f"Excluding: {len(exclude)} / {len(motifs_df)} rows")
motifs_df['is_erv'] = exclude_rows #.astype(str)

#check motif counts without ERVs
dfi = motifs_df[(motifs_df['is_erv'].astype(str)=='False')]
dfi.is_erv.value_counts()

# above returened dataframe without erv, I will make pyrnage to resize the motif to 18 bp from start and make df again for downstream analysis
motif_pr = pr.PyRanges(dfi)
motif_pr = motif_pr.extend({'3':1})
x =motif_pr.df
x

#Get sequences on the forward strand
x['fwd_seq'] = extract_seqs_from_df(coords_df = x, fasta_path = fasta_file, chrom_column = 'Chromosome', start_column = 'Start', end_column = 'End')
x

# Reorient the negative strand sequences
motifs_oriented_df = pd.DataFrame()
for i,row in tqdm(x.iterrows()):
    if row.strand=='-':
        fwd_seq = Seq(row.fwd_seq)
        rev_seq = str(fwd_seq.reverse_complement())
        row['seq'] = rev_seq
    else:
        row['seq'] = row.fwd_seq
    motifs_oriented_df = motifs_oriented_df.append(row)

## Extract first half site and second half site groups.

first_half_site = (0,9) 
second_half_site = (9,19) 

motifs_oriented_df['first_half_site_seq'] = [s[first_half_site[0]:first_half_site[1]] for s in motifs_oriented_df.seq.values]
motifs_oriented_df['second_half_site_seq'] = [s[second_half_site[0]:second_half_site[1]] for s in motifs_oriented_df.seq.values]

#Save results.
motifs_oriented_df.to_csv('tsv/cwm_tead4double_noervs_nopromoter_withseq.tsv.gz', sep = '\t', index = False)

motifs_oriented_df

## Group results based on half site summaries.

summarized_tead4_sites_df = motifs_oriented_df.groupby(['first_half_site_seq', 'second_half_site_seq']).size().reset_index()
summarized_tead4_sites_df.columns = ['first_half_site_seq', 'second_half_site_seq', 'frequency']
summarized_tead4_sites_df = summarized_tead4_sites_df.sort_values('frequency', ascending = False)
summarized_tead4_sites_df = summarized_tead4_sites_df.reset_index()

summarized_tead4_sites_df1=summarized_tead4_sites_df[0:5000]

summarized_tead4_sites_df1

summarized_tead4_sites_df.shape
summarized_tead4_sites_df

## Get in silico contributions for each combination of half sites. 

### A. For each combination...
    1. Generate random sequences.
    2. Inject each half site at appropriate distances.
    3. Get predict_all across all trials
    4. Average across appropriate areas for contribution.

## functions for generating random sequences & encoding and decoding in 1he format

# seq in form of ACGT letters
def generate_random_seq(seqlen, weights = [.25, .25, .25, .25]):
    """
    Purpose: Generate a random DNA sequence of a specified length.
    """
    import random
    seq = random.choices(['A','C','G','T'], weights = weights, k=seqlen)
    return(''.join(seq))

# 1he seq
def one_hot_encode_sequences(sequences):
    """
    Purpose: Given an array of sequences, one-hot-encode into a [region x position x 4] array.
    """
    return(np.stack([one_hot_encode_sequence(s) for s in sequences]))


def one_hot_encode_sequence(sequence):
    """
    Kudos to Charles: /n/projects/cm2363/bpnet-nucleosomes/work/localimportance/allLocalImportances.py
    Purpose: Given a SINGLE sequence string, one-hot encode the data.
        + default control_profiles and control_logcounts is to be zeroed out
        + naively detects whether the sequence is one-hot-encoded.
    """
    onehot_mapping = {
    'A': [1,0,0,0],
    'C': [0,1,0,0],
    'G': [0,0,1,0],
    'T': [0,0,0,1],
    'a': [1,0,0,0],
    'c': [0,1,0,0],
    'g': [0,0,1,0],
    't': [0,0,0,1],
    'N': [0,0,0,0]
    }
    return np.array([onehot_mapping[x] for x in sequence])


def one_hot_decode_sequence(array):
    """
    Purpose: Given an array [position x 4], decode sequence to a string.
    """
    onehot_decoder = {
    0: 'A',
    1: 'C',
    2: 'G',
    3: 'T'
    }

    idxs = np.where(array)[1]
    return (''.join([onehot_decoder[i] for i in idxs]))

# generate 64 random seq
import random
random.seed(10)
trials = 64
random_seqs = [generate_random_seq(seqlen = 1000) for i in range(trials)]

# make it 1he
random_seqs_1he =one_hot_encode_sequences(random_seqs)
flanks = (random_seqs_1he.shape[1]-1000)//2
random_seqs_1he = random_seqs_1he[:, flanks:(1000+flanks)]
random_seqs_1he.shape

## Collect all sequences with a pd.df that keeps indexes.

combination_index_df = pd.DataFrame()
combination_w_trials_list = []

for i,row in tqdm(summarized_tead4_sites_df.iterrows()):
        
    #define half sites
    first_half_site = row.first_half_site_seq
    second_half_site = row.second_half_site_seq
    
    #Create sequences
    for j,s in enumerate(random_seqs_1he):
        seq = one_hot_decode_sequence(s)
        seq_w_first = insert_motif(seq = seq, motif = first_half_site, position = 500)
        seq_w_second = insert_motif(seq = seq_w_first, motif = second_half_site, position = 509)
        combination_w_trials_list.append(seq_w_second)
        row['trial_index'] = j
        row['combo_index'] = i
        combination_index_df = combination_index_df.append(row)

#Create row index that combines trial index and ocmbo index and matches index of contrib predictions
combination_index_df['row_index'] = (combination_index_df.combo_index * 64 + combination_index_df.trial_index).astype(int)

len(combination_w_trials_list)

#to save the dataframe
#combination_index_df.to_pickle('tsv/tsc_tead4double_combination_index_df.pkl')    
#combination_index_df = pd.read_pickle('tsv/tsc_tead4double_combination_index_df.pkl')

##### In a single prediction step, get all contribution scores (`list(3K) -> ['contrib_score']['tead4/profile']`), save as .pkl file.

%%script false --no-raise-error
import pickle

#One hot encode sequences
combination_w_trixals_seq_1he = one_hot_encode_sequences(combination_w_trials_list)

#Predict all sequences
combinations_w_trials_pred_dict = model.predict_all(combination_w_trixals_seq_1he, contrib_method='deeplift')
with open('tsv/all_combination_of_halfsites_w_trials.pkl', 'wb') as f:
    pickle.dump(combinations_w_trials_pred_dict, f)

### Import pkl file and isolate and average values. 

import pickle
with open('tsv/all_combination_of_halfsites_w_trials.pkl', 'rb') as f:
    combinations_w_trials_pred_dict = pickle.load(f)

summarized_tead4_sites_w_contrib_df = pd.DataFrame()

for i in tqdm(range(len(combinations_w_trials_pred_dict))):
     #Collect contribution scores under proper coordinates for all trials
    first_half_site_contrib = (combinations_w_trials_pred_dict[i]['contrib_score']['tead4/profile'][496:505] * combinations_w_trials_pred_dict[i]['seq'][496:505]).sum()
    second_half_site_contrib = (combinations_w_trials_pred_dict[i]['contrib_score']['tead4/profile'][505:514] * combinations_w_trials_pred_dict[i]['seq'][505:514]).sum()
    df = pd.DataFrame()
    df['first_half_site_contrib'] = [first_half_site_contrib]
    df['second_half_site_contrib'] = [second_half_site_contrib]
    df['row_index'] = i

    summarized_tead4_sites_w_contrib_df = summarized_tead4_sites_w_contrib_df.append(df) 

## Reset index and average the contribution over trial

contrib_over_combo_over_trials_df = combination_index_df.reset_index().join(summarized_tead4_sites_w_contrib_df.reset_index(), on = ['row_index'], how = 'left', rsuffix = 'test')
contrib_over_combo_df = contrib_over_combo_over_trials_df.groupby(['combo_index', 'first_half_site_seq', 'second_half_site_seq']).agg({'first_half_site_contrib' : 'mean', 'second_half_site_contrib' : 'mean'}).reset_index()
contrib_over_combo_df

### Cut the contrib values in each half into quantiles and save

first_half_site_cutoffs = np.quantile(contrib_over_combo_df.first_half_site_contrib.values, [0, .25, .5, .75, 1])
second_half_site_cutoffs = np.quantile(contrib_over_combo_df.second_half_site_contrib.values, [0, .25, .5, .75, 1])

contrib_over_combo_df['first_half_site_q'] = pd.cut(contrib_over_combo_df.first_half_site_contrib, bins=first_half_site_cutoffs, labels=['q1','q2','q3','q4'], include_lowest = True)
contrib_over_combo_df['second_half_site_q'] = pd.cut(contrib_over_combo_df.second_half_site_contrib, bins=second_half_site_cutoffs, labels=['q1','q2','q3','q4'], include_lowest = True)

#save
contrib_over_combo_df.to_csv('tsv/tead4_double_motif_contrib_over_combo_completehalfs_quantiles.tsv.gz', sep = '\t', index = False)

#check freq
df = contrib_over_combo_df.value_counts(['first_half_site_q', 'second_half_site_q']).reset_index()


## Get insilico predictions keeping the combo index from above df

1. get random seq decoded as list
2. get first_half_site & second half sites from contrib_over_combo_df, keeping the combo index.
3. inject each tead4 combo and measure total counts with and without entire motif (background).
4. append to df above and save the df for plotting in R

random_seqs_1he.shape

## get random seq decoded as list
seqs = []
for j,s in enumerate(random_seqs_1he):
    seqs.append(one_hot_decode_sequence(s))

def collect_summarized_counts(contrib_over_combo_df, random_seqs, tasks_of_interest, second_motif_injection_site, motif_window, measurement_window):

    """
    Purpose: Given WT, dA, dB, and dAdB states of a WT genomic tead4 halfs -> randomize other halfs, how does the double motif enrichment change?
    Inputs:
    Outputs:
    """
    
    from tqdm import tqdm
    from concise.preprocessing import encodeDNA
    def generate_seq(random_seq, central_motif, side_motif=None, side_distances=[], seqlen=1000):
        from bpnet.simulate import insert_motif
        injected_seq = insert_motif(random_seq, central_motif, seqlen // 2)
        for d in side_distances:
            injected_seq = insert_motif(injected_seq, side_motif, d)
        return injected_seq
    
    summary_for_all_combos_df = pd.DataFrame()
    for i,row in tqdm(contrib_over_combo_df.iterrows()):
        
        #define half sites
        first_half_site = row.first_half_site_seq
        second_half_site = row.second_half_site_seq
        combo_index = row.combo_index

        #Generate sequences with variant states
        WT_seqs = [generate_seq(random_seq = i, central_motif = first_half_site, side_motif = second_half_site, 
                                side_distances = [second_motif_injection_site]) for i in random_seqs]
        dAdB_seqs = [generate_seq(random_seq = i, central_motif = '', side_motif = '', 
                                  side_distances = []) for i in random_seqs]
        dA_seqs = [generate_seq(random_seq = i, central_motif = '', side_motif = second_half_site, 
                                side_distances = [second_motif_injection_site]) for i in random_seqs]
        dB_seqs = [generate_seq(random_seq = i, central_motif = first_half_site, side_motif = '', 
                                side_distances = [second_motif_injection_site]) for i in random_seqs]

        #Collect into dictionary
        seq_dict = {'WT': WT_seqs, 'dAdB': dAdB_seqs, 'dA': dA_seqs, 'dB': dB_seqs}


        #Predict each variant state({inj_state -> {task -> [64 x 1000 x 2]}))
        preds_dict = {k: model.predict(encodeDNA(v)) for k,v in seq_dict.items()}
        
        # Average across trials and measurement window({inj_state -> {task -> #}})
        preds_single_dict = {inj_state: {task: np.mean(np.sum(v1[:, measurement_window[0]:measurement_window[1], :], axis = (1,2))) 
                     for task,v1 in v.items()} 
         for inj_state,v in preds_dict.items()}

        #Summarize counts together pd.df based on predictions
        summary_df = pd.DataFrame()
        for inj_state, v in preds_single_dict.items():
            df = pd.DataFrame.from_dict(v, orient = 'index').transpose()
            df['inj_state'] = inj_state
            df = df.melt(id_vars = ['inj_state'], var_name = 'task', value_name = 'counts')
            summary_df = summary_df.append(df)
        summary_df['first_half_site'] = first_half_site
        summary_df['second_half_site'] = second_half_site        
        summary_df['combo_index'] = combo_index
        summary_for_all_combos_df = summary_for_all_combos_df.append(summary_df)
    return(summary_for_all_combos_df)

## Collect summarized counts and profiles across trials for each halfs.

counts_df = collect_summarized_counts(contrib_over_combo_df=contrib_over_combo_df,random_seqs = seqs, 
                                   tasks_of_interest = ['tead4'], #','yap1'
                                   second_motif_injection_site = 509,
                                   motif_window = (496, 514), measurement_window = (475, 535))
#counts_df

#save .tsv.gz for each object
counts_df.to_csv(f'csv/all_counts_summary_predictions_genomic_regions_9mer.csv.gz', index = False)
