==> bpnet/analysis/8a_getting_contrib_for_tead_patterns.ipynb <==
# Introduction

Get contribution score for tead single and tead double motif sequences patterns that were part of affinity plot in analysis 8_

# Computational setup

## Environment

import warnings
warnings.filterwarnings("ignore")
from tensorflow.python.util import deprecation
deprecation._PRINT_DEPRECATION_WARNINGS = False

#Packages
import os
import sys
import pandas as pd
import numpy as np
import plotnine
import seaborn as sns
from pybedtools import BedTool
from tqdm import tqdm
from bpnet.utils import create_tf_session
from bpnet.BPNet import BPNetSeqModel
from bpnet.simulate import motif_coords
from bpnet.preproc import resize_interval
from concise.preprocessing import encodeDNA
from bpnet.extractors import extract_seq
from plotnine import *
from Bio.Seq import Seq

# Settings
os.chdir(f'/n/projects/kd2200/publication/bpnet/analysis/')
pd.set_option('display.max_columns', 100)
create_tf_session('0',1)

# Custom functions
sys.path.insert(0, f'/n/projects/kd2200/publication/bpnet/analysis/scripts/bpnet/scripts/')
from perturb_functions import insert_motif, one_hot_decode_sequence, one_hot_encode_sequences
from motif_functions import remove_palindromic_motif_duplicates
from data_format_functions import tidy_bpnet_predictions_nexus,tidy_bpnet_contributions
from motifs import extract_seqs_from_df


## Variables

%matplotlib inline

# Pre-existing variables
fasta_file = f'/n/projects/kd2200/publication/bpnet/fasta/mm10.fa'
model_dir = f'/n/projects/kd2200/publication/bpnet/model/dataspec.yaml_default_fold_5/'

# Independent variables
#tead double motif patterns
#tead4_motifs_path = f'tsv/tdbl_motifs_overlap_with_affinity_seq_1based.tsv.gz'

#tead single motif patterns
tead4_motifs_path = f'tsv/td4_single_motifs_overlap_with_affinity_seq_1based.tsv.gz'

## Load BPNet model

model = BPNetSeqModel.from_mdir(model_dir)

# Analysis code
## Import Tead single or double motifs (one pattern set at a time)

motifs_df = pd.read_csv(tead4_motifs_path, sep = '\t')
motifs_df['seq'] = motifs_df['sequence']
motifs_df

## denote coordinate based on eith Tead single or double motifs

#tead double motif patterns
#sites = (0,17)
#motifs_df['sites_seq'] = [s[sites[0]:sites[1]] for s in motifs_df.seq.values]

#tead single motif patterns
sites = (0,10)
motifs_df['sites_seq'] = [s[sites[0]:sites[1]] for s in motifs_df.seq.values]

motifs_df

## Group results based on pattern match summaries.

summarized_tead4_sites_df = motifs_df.groupby(['sites_seq']).size().reset_index()
summarized_tead4_sites_df.columns = ['sites_seq', 'frequency']
summarized_tead4_sites_df = summarized_tead4_sites_df.sort_values('frequency', ascending = False)
summarized_tead4_sites_df = summarized_tead4_sites_df.reset_index()
summarized_tead4_sites_df

## Get contributions in silico for motifs. 

1. For each motifs
2. Import random sequences.
3. Inject each motifs in random seq.
4. Get predict_all across all trials
5. Average contribution across appropriate areas for trial for all the motifs

trials = 64
random_seqs = np.load('/n/projects/mw2098/analysis/chrombpnet/models/atac/mm10/data/pkl/random_seqs_seed_10_trials_256_array.npz')['seqs_1he'][:trials]
flanks = (random_seqs.shape[1]-1000)//2
random_seqs = random_seqs[:, flanks:(1000+flanks)]
random_seqs.shape

## functions for generating random sequences & encoding and decoding in 1he format

# seq in form of ACGT letters
def generate_random_seq(seqlen, weights = [.25, .25, .25, .25]):
    """
    Purpose: Generate a random DNA sequence of a specified length.
    """
    import random
    seq = random.choices(['A','C','G','T'], weights = weights, k=seqlen)
    return(''.join(seq))

# 1he seq
def one_hot_encode_sequences(sequences):
    """
    Purpose: Given an array of sequences, one-hot-encode into a [region x position x 4] array.
    """
    return(np.stack([one_hot_encode_sequence(s) for s in sequences]))


def one_hot_encode_sequence(sequence):
    """
    Kudos to Charles: /n/projects/cm2363/bpnet-nucleosomes/work/localimportance/allLocalImportances.py
    Purpose: Given a SINGLE sequence string, one-hot encode the data.
        + default control_profiles and control_logcounts is to be zeroed out
        + naively detects whether the sequence is one-hot-encoded.
    """
    onehot_mapping = {
    'A': [1,0,0,0],
    'C': [0,1,0,0],
    'G': [0,0,1,0],
    'T': [0,0,0,1],
    'a': [1,0,0,0],
    'c': [0,1,0,0],
    'g': [0,0,1,0],
    't': [0,0,0,1],
    'N': [0,0,0,0]
    }
    return np.array([onehot_mapping[x] for x in sequence])


def one_hot_decode_sequence(array):
    """
    Purpose: Given an array [position x 4], decode sequence to a string.
    """
    onehot_decoder = {
    0: 'A',
    1: 'C',
    2: 'G',
    3: 'T'
    }

    idxs = np.where(array)[1]
    return (''.join([onehot_decoder[i] for i in idxs]))

# generate 64 random seq
import random
random.seed(10)
trials = 64
random_seqs = [generate_random_seq(seqlen = 1000) for i in range(trials)]

# make it 1he
random_seqs_1he =one_hot_encode_sequences(random_seqs)
flanks = (random_seqs_1he.shape[1]-1000)//2
random_seqs_1he = random_seqs_1he[:, flanks:(1000+flanks)]
random_seqs_1he.shape

combination_index_df = pd.DataFrame()
combination_w_trials_list = []

for i,row in tqdm(summarized_tead4_sites_df.iterrows()):
        
    #define half sites
    sites = row.sites_seq
    
    #Create sequences
    for j,s in enumerate(random_seqs_1he):
        seq = one_hot_decode_sequence(s)
        seq_w_motif = insert_motif(seq = seq, motif = sites, position = 500)
        combination_w_trials_list.append(seq_w_motif)
        row['trial_index'] = j
        row['combo_index'] = i
        combination_index_df = combination_index_df.append(row)

#Create row index that combines trial index and ocmbo index and matches index of contrib predictions
combination_index_df['row_index'] = (combination_index_df.combo_index * 64 + combination_index_df.trial_index).astype(int)

len(combination_w_trials_list)

%%script false --no-raise-error

import pickle

#One hot encode sequences
combination_w_trials_seq_1he = one_hot_encode_sequences(combination_w_trials_list)

#Predict all sequences
combinations_w_trials_pred_dict = model.predict_all(combination_w_trials_seq_1he, contrib_method='deeplift')

with open('tsv/td4_single_motifs_overlap_with_affinity_seq.pkl', 'wb') as f:
    pickle.dump(combinations_w_trials_pred_dict, f)
    

# import pkl file

# tead double motif patterns
# import pickle
# with open('tsv/tdbl_motifs_overlap_with_affinity_seq.pkl', 'rb') as f:
#     combinations_w_trials_pred_dict = pickle.load(f)

# tead single motif patterns 
import pickle
with open('tsv/td4_single_motifs_overlap_with_affinity_seq.pkl', 'rb') as f:
    combinations_w_trials_pred_dict = pickle.load(f)


def one_hot_decode_sequence(array):
    """
    Purpose: Given an array [position x 4], decode sequence to a string.
    """
    onehot_decoder = {
    0: 'A',
    1: 'C',
    2: 'G',
    3: 'T'
    }

    idxs = np.where(array)[1]
    return (''.join([onehot_decoder[i] for i in idxs]))


#sanity check for tead4 double
# xx =combinations_w_trials_pred_dict[1]['contrib_score']['tead4/profile'][492:509]* combinations_w_trials_pred_dict[1]['seq'][492:509]
# DF = pd.DataFrame(xx)
# DF

#sanity check for tead4 single
xx =combinations_w_trials_pred_dict[1]['contrib_score']['tead4/profile'][496:505]* combinations_w_trials_pred_dict[1]['seq'][496:505]
DF = pd.DataFrame(xx)
DF

# for tead4 double
# contrib_by_seq = []
# for j,i in tqdm(enumerate(range(0, len(combinations_w_trials_pred_dict), trials))):
#     contrib_by_trial = []
#     for k in range(trials):
#          #Collect contribution scores under proper coordinates for all trials 
#         contrib = [(combinations_w_trials_pred_dict[i+k]['contrib_score']['tead4/profile'][492:509] * combinations_w_trials_pred_dict[i+k]['seq'][492:509])]
#         contrib_by_trial.append(contrib)
#     contrib_avg = np.mean(np.squeeze(np.array(contrib_by_trial)), axis = 0)
#     contrib_by_seq.append([contrib_avg])
# contrib_by_seqs = np.squeeze(np.array(contrib_by_seq))
# cwm = np.mean(contrib_by_seqs, axis = 0)

# cwm_for_plotting =pd.DataFrame(cwm, columns = ['A','C','G','T']).transpose()


# for tead4 single
contrib_by_seq = []
for j,i in tqdm(enumerate(range(0, len(combinations_w_trials_pred_dict), trials))):
    contrib_by_trial = []
    for k in range(trials):
         #Collect contribution scores under proper coordinates for all trials 
        contrib = [(combinations_w_trials_pred_dict[i+k]['contrib_score']['tead4/profile'][496:505] * combinations_w_trials_pred_dict[i+k]['seq'][496:505])]
        contrib_by_trial.append(contrib)
    contrib_avg = np.mean(np.squeeze(np.array(contrib_by_trial)), axis = 0)
    contrib_by_seq.append([contrib_avg])
contrib_by_seqs = np.squeeze(np.array(contrib_by_seq))
cwm = np.mean(contrib_by_seqs, axis = 0)

cwm_for_plotting =pd.DataFrame(cwm, columns = ['A','C','G','T']).transpose()

contrib_by_seq[0]

contrib_by_seq[1]

#saved cwm matrix as tsv

#cwm_for_plotting.to_csv('tsv/tead4_double_motif_contrib_forpatterns.tsv', sep = '\t', index = False)
cwm_for_plotting.to_csv('tsv/tead4_single_motif_contrib_forpatterns.tsv', sep = '\t', index = False)

#saved summary as csv

#summarized_tead4_sites_df.to_csv('tsv/tead4_double_motif_freq_forpatterns.csv', sep = '\t', index = False)
summarized_tead4_sites_df.to_csv('tsv/tead4_single_motif_freq_forpatterns.csv', sep = '\t', index = False)

## Below code was run in R using library(ggseqlogo)

#import tsv
tdbl <- read_tsv("tsv/tead4_double_motif_contrib_forpatterns.tsv",col_names = TRUE)
td4 <- read_tsv("tsv/tead4_single_motif_contrib_forpatterns.tsv",col_names = TRUE)

#make df
tdbl <- as.data.frame(tdbl)
td4 <- as.data.frame(td4)

#make matrix
#rownames(tdbl)<- c("A","C","G","T")
#rownames(td4)<- c("A","C","G","T")
tdbl <- as.matrix.data.frame(tdbl)
td4 <- as.matrix.data.frame(td4)

#plot
g_tdbl <- ggseqlogo(tdbl, method='custom', seq_type='dna',scales = 'none') + ylab('CWM')
g_td4 <- ggseqlogo(td4, method='custom', seq_type='dna',scales = 'none') + ylab('CWM')

# save plot
ggsave("figures/9_tead4_double_motif_analysis/cwm_tead4_double_patterns_contrib.pdf",g_tdbl, height = 4, width = 6)
ggsave("figures/9_tead4_double_motif_analysis/cwm_tead4_single_patterns_contrib.pdf",g_td4, height = 4, width = 6)

cwm_for_plotting
