==> bpnet/analysis/15_generate_prediction_forluci_assay.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Introduction" data-toc-modified-id="Introduction-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Introduction</a></span></li><li><span><a href="#BPNet-Setup" data-toc-modified-id="BPNet-Setup-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>BPNet Setup</a></span><ul class="toc-item"><li><span><a href="#Inputs:" data-toc-modified-id="Inputs:-2.1"><span class="toc-item-num">2.1&nbsp;&nbsp;</span>Inputs:</a></span></li></ul></li><li><span><a href="#Import-data-needed" data-toc-modified-id="Import-data-needed-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Import data needed</a></span><ul class="toc-item"><li><span><a href="#Generate-the-importance-scores" data-toc-modified-id="Generate-the-importance-scores-3.1"><span class="toc-item-num">3.1&nbsp;&nbsp;</span>Generate the importance scores</a></span></li></ul></li><li><span><a href="#Generate-predictions" data-toc-modified-id="Generate-predictions-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Generate predictions</a></span><ul class="toc-item"><li><span><a href="#Reformat-predictions" data-toc-modified-id="Reformat-predictions-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span>Reformat predictions</a></span></li></ul></li><li><span><a href="#Export-predictions" data-toc-modified-id="Export-predictions-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Export predictions</a></span></li></ul></div>

# Introduction

We want to generate prediction profiles and importance scores for given plasmid sequences based on the trained a model in mtsc cells.


# BPNet Setup



print("Launching packages...")
import pandas as pd 
import os
import json
import numpy as np
import comet_ml
from comet_ml import Experiment
from collections import OrderedDict

import os
import sys
import pandas as pd
import numpy as np
from pybedtools import BedTool
from tqdm import tqdm
from bpnet.cli.contrib import bpnet_contrib
from bpnet.cli.modisco import cwm_scan
from bpnet.cli.contrib import ContribFile
from bpnet.BPNet import BPNetSeqModel
from concise.preprocessing import encodeDNA
from bpnet.simulate import random_seq, insert_motif
from bpnet.plot.tracks import plot_tracks, to_neg

from pathlib import Path
from concise.preprocessing import encodeDNA
from concise.preprocessing.sequence import one_hot2string, DNA
from bpnet.BPNet import BPNetSeqModel
import warnings
warnings.filterwarnings("ignore")

## Inputs:

#Set wd 
os.chdir(f'/n//projects/kd2200/analysis/BPNet/cegkttyz_tsc_bpnet/grid_search/final_cttgy_tsc_model/bpnet/analysis/fold_5/csv/for_preds/predictions/')


#Set model directory 

mdir = f'/n//projects/kd2200/analysis/BPNet/cegkttyz_tsc_bpnet/grid_search/final_cttgy_tsc_model/bpnet/model/dataspec.yaml_default_fold_5/'


#bpnet = BPNetPredictor.from_mdir(mdir)
model = BPNetSeqModel.from_mdir(mdir)

# Import data needed


print("Reading sequences...")
can_cited2_input=pd.read_csv("../can_cited2_1000bp.csv") 
can_amotl2_input=pd.read_csv("../can_amotl2_1000bp.csv") 
can_bmp7_input=pd.read_csv("../can_bmp7_1000bp_tty.csv") 
can_tjp1_input=pd.read_csv("../can_tjp1_1000bp_negstrand.csv") 
can_rin3_input=pd.read_csv("../can_rin3_2bpdel_1000bp.csv") 
can_ezr_input=pd.read_csv("../can_ezr_1000bp.csv") 
can_tjp1_input.head(n=10)

## Generate the importance scores 
Apply the trained Oct4/Sox2/Klf4/Nanog model to predict importance scores.

1. 1-hot encode the data.  
2. Predict the ChIP-nexus profiles using BPNet 

#1-hot-encode the data
can_cited2 = encodeDNA(can_cited2_input.sequence.str.upper())
can_amotl2 = encodeDNA(can_amotl2_input.sequence.str.upper())
can_bmp7 = encodeDNA(can_bmp7_input.sequence.str.upper())
can_tjp1 = encodeDNA(can_tjp1_input.sequence.str.upper())
can_rin3 = encodeDNA(can_rin3_input.sequence.str.upper())
can_ezr = encodeDNA(can_ezr_input.sequence.str.upper())

# Generate predictions

#Predict the data
can_cited2_seq_preds = model.predict_all(can_cited2, contrib_method='deeplift')
can_amotl2_seq_preds = model.predict_all(can_amotl2, contrib_method='deeplift')
can_bmp7_seq_preds = model.predict_all(can_bmp7, contrib_method='deeplift')
can_tjp1_seq_preds = model.predict_all(can_tjp1, contrib_method='deeplift')
can_rin3_seq_preds = model.predict_all(can_rin3, contrib_method='deeplift')
can_ezr_seq_preds = model.predict_all(can_ezr, contrib_method='deeplift')

## Reformat predictions 

#from basepair.plot.tracks import plot_tracks, filter_tracks
#from basepair.plot.profiles import plot_stranded_profile
def bpnet_format_dict_to_df(dict_obj):
    tead4_prediction = pd.DataFrame(data= dict_obj['p']["tead4"], columns = ["pos", "neg"])
    tfap2c_prediction = pd.DataFrame(data= dict_obj['p']["tfap2c"], columns = ["pos", "neg"])
    cdx2_prediction = pd.DataFrame(data= dict_obj['p']["cdx2"], columns = ["pos", "neg"])
    gata3_prediction = pd.DataFrame(data= dict_obj['p']["gata3"], columns = ["pos", "neg"])
    yap1_prediction = pd.DataFrame(data= dict_obj['p']["yap1"], columns = ["pos", "neg"])

    tead4_importance = pd.DataFrame(data= dict_obj['i']["tead4"], columns = ["score"])
    tfap2c_importance = pd.DataFrame(data= dict_obj['i']["tfap2c"], columns = ["score"])
    cdx2_importance = pd.DataFrame(data= dict_obj['i']["cdx2"], columns = ["score"])
    gata3_importance = pd.DataFrame(data= dict_obj['i']["gata3"], columns = ["score"])
    yap1_importance = pd.DataFrame(data= dict_obj['i']["yap1"], columns = ["score"])

    tead4_data = pd.concat([tead4_prediction, tead4_importance], axis=1)
    tfap2c_data = pd.concat([tfap2c_prediction, tfap2c_importance], axis=1)
    cdx2_data = pd.concat([cdx2_prediction, cdx2_importance], axis=1)
    gata3_data = pd.concat([gata3_prediction, gata3_importance], axis=1)
    yap1_data = pd.concat([yap1_prediction, yap1_importance], axis=1)

    tead4_data["factor"]='Tead4'
    tfap2c_data["factor"]='Tfap2c'
    cdx2_data["factor"]='Cdx2'
    gata3_data["factor"]='Gata3'
    yap1_data["factor"]='Yap1'

    tead4_data['position']=range(tead4_data.shape[0])
    tfap2c_data['position']=range(tfap2c_data.shape[0])
    cdx2_data['position']=range(cdx2_data.shape[0])
    gata3_data['position']=range(gata3_data.shape[0])
    yap1_data['position']=range(yap1_data.shape[0])


    output_data = pd.concat([tead4_data, tfap2c_data, cdx2_data,gata3_data, yap1_data], axis=0) 
    return(output_data)

def bpnet_get_dict_output(obj, tasks):
    out = dict(p=obj['preds'],
               i={t: obj['seq'] * (obj['contrib_score'][f'{t}/profile']['pos'] + obj['contrib_score'][f'{t}/profile']['neg'])/2 
                  for ti,t in enumerate(tasks)})    
    for i in range(len(out['i'])):
        out['i'][bpnet.tasks[i]] = np.sum(out['i'][bpnet.tasks[i]], axis=1).tolist() #add quick loop to format seqs
        
    output_df=bpnet_format_dict_to_df(out)
    
    return(output_df)

# Export predictions

def export_predictions_all(preds_all, tasks, contrib_type = 'profile'):
    import sys
    sys.path.insert(0, f'/n/projects/mw2098/shared_code/bpnet/scripts')
    from data_format_functions import tidy_bpnet_predictions_nexus 
    
    #Create empty pd.df to write into 
    preds_df = pd.DataFrame()
    contrib_df = pd.DataFrame() 
    
    #Tidy perturbations into a pd.df
    for task in tasks:
        for r in range(len(preds_all)):
            p = {task: preds_all[r]['pred'][task]}
            p_df = tidy_bpnet_predictions_nexus(predictions = p, tasks = [task])
            p_df['task'] = task
            p_df['region_idx'] = r
            preds_df = preds_df.append(p_df) 
            c_df = pd.DataFrame(preds_all[r]['contrib_score'][f'{task}/{contrib_type}']*preds_all[r]['seq'], columns = ['A','C','G','T'])
            #c_df = pd.DataFrame(preds_all[r]['contrib_score'][f'{task}/{contrib_type}'], columns = ['A','C','G','T'])
            c_df['task'] = task
            c_df['position'] = list(range(c_df.shape[0]))
            c_df['region_idx'] = r
            contrib_df = contrib_df.append(c_df)
    return(preds_df, contrib_df)
    

can_cited2_predictions, can_cited2_contribution = export_predictions_all(preds_all=can_cited2_seq_preds, tasks=['tead4', 'tfap2c', 'yap1', 'cdx2', 'gata3']) 
can_amotl2_predictions, can_amotl2_contribution = export_predictions_all(preds_all=can_amotl2_seq_preds, tasks=['tead4', 'tfap2c', 'yap1', 'cdx2', 'gata3']) 
can_bmp7_predictions, can_bmp7_contribution = export_predictions_all(preds_all=can_bmp7_seq_preds, tasks=['tead4', 'tfap2c', 'yap1', 'cdx2', 'gata3']) 
can_tjp1_predictions, can_tjp1_contribution = export_predictions_all(preds_all=can_tjp1_seq_preds, tasks=['tead4', 'tfap2c', 'yap1', 'cdx2', 'gata3']) 
can_rin3_predictions, can_rin3_contribution = export_predictions_all(preds_all=can_rin3_seq_preds, tasks=['tead4', 'tfap2c', 'yap1', 'cdx2', 'gata3']) 
can_ezr_predictions, can_ezr_contribution = export_predictions_all(preds_all=can_ezr_seq_preds, tasks=['tead4', 'tfap2c', 'yap1', 'cdx2', 'gata3']) 


can_ezr_predictions

can_ezr_contribution

#Write to csv for import to R
can_cited2_predictions.to_csv("can_cited2_predict.csv",index = False)
can_cited2_contribution.to_csv("can_cited2_contrib.csv",index = False)

can_amotl2_predictions.to_csv("can_amotl2_predict.csv",index = False)
can_amotl2_contribution.to_csv("can_amotl2_contrib.csv",index = False)

can_bmp7_predictions.to_csv("can_bmp7_predict.csv",index = False)
can_bmp7_contribution.to_csv("can_bmp7_contrib.csv",index = False)

can_tjp1_predictions.to_csv("can_tjp1_predict.csv",index = False)
can_tjp1_contribution.to_csv("can_tjp1_contrib.csv",index = False)

can_tjp1_predictions.to_csv("can_tjp1_predict_negstrand.csv",index = False)
can_tjp1_contribution.to_csv("can_tjp1_contrib_negstrand.csv",index = False)

can_rin3_predictions.to_csv("can_rin3_predict.csv",index = False)
can_rin3_contribution.to_csv("can_rin3_contrib.csv",index = False)

can_ezr_predictions.to_csv("can_ezr_predict.csv",index = False)
can_ezr_contribution.to_csv("can_ezr_contrib.csv",index = False)
