==> bpnet/analysis/18_distance_perturbation.ipynb <==
# Introduction

The purpose of this `.ipynb` is to use the `TSC TFs tranied BPNet model` to predict binding on genomic regions where Tfap2c and Tead4 have distance replationship and plot change in predicted binding upon their distance manipulation.

# Computational setup

import warnings
warnings.filterwarnings("ignore")
from tensorflow.python.util import deprecation
deprecation._PRINT_DEPRECATION_WARNINGS = False

#Packages
import os
import sys
import pandas as pd
import numpy as np
from pybedtools import BedTool
from genomelake.extractors import FastaExtractor
from bpnet.extractors import Interval
from bpnet.cli.contrib import bpnet_contrib
from bpnet.cli.modisco import cwm_scan
from bpnet.cli.contrib import ContribFile
from concise.preprocessing import encodeDNA
from bpnet.cli.contrib import ContribFile
from bpnet.simulate import random_seq, insert_motif
from bpnet.plot.tracks import plot_tracks, to_neg
from plotnine import *
from tqdm import tqdm
from bpnet.BPNet import BPNetSeqModel
from bpnet.utils import create_tf_session

#create_tf_session('0', .4)

# Settings
os.chdir(f'/n/projects/kd2200/publication/bpnet/analysis/test/')
pd.set_option('display.max_columns', 100)


# Custom commands
sys.path.insert(0, f'/n/projects/kd2200/publication/bpnet/analysis/scripts/bpnet/scripts')
from data_format_functions import myround, myfloor, myceiling, df_to_intervals,tidy_bpnet_predictions_nexus, one_hot_decode
from motif_functions import remove_palindromic_motif_duplicates, filter_overlapping_motifs_by_priority


# function to return key for any value 
def get_key(val, my_dict): 
    for key, value in my_dict.items(): 
        if val == value: 
            return key 
    return "key doesn't exist"

#Function to save contribution information with 4 columns and a task (does not use hypoth. contrib)
def tidy_bpnet_contributions(seq, contrib, tasks, contrib_type = 'profile'):
    contrib_df = pd.DataFrame()
    for task in tasks:
        c = contrib[f'{task}/{contrib_type}'] * seq
        df = pd.DataFrame(c, columns = ['A','C','G','T'])
        df['task'] = task
        contrib_df=contrib_df.append(df)
    return(contrib_df)

#Pre-existing variables
fasta_file = f'/n/projects/kd2200/publication/bpnet/fasta/mm10.fa'
model_dir = f'/n/projects/kd2200/publication/bpnet/model/dataspec.yaml_default_fold_5/'
contrib_file = f"/n/projects/kd2200/publication/bpnet/contrib/dataspec.yaml_default_fold_5/contrib.h5"
modisco_dir = f'/n/projects/kd2200/publication/bpnet/modisco/profile/dataspec.yaml_default_fold_5/'
modisco_all_region_cwm_dir = f'/n/projects/kd2200/publication/bpnet/cwm_all_regions/'
tasks = ['tead4','tfap2c','yap1','gata3','cdx2']

# Independent variables
motifs_of_interest_dict = {'tead4': 'tead4/m0_p0',
                           'tfap2c': 'tfap2c/m0_p0',
                            'cdx2': 'cdx2/m0_p0',
                            'gata3': 'gata3/m0_p4'}
# Dependent variables
enhancers_path = f'enhancers_genes_tead4_tfap2c_dis_0based.bed'
enhancers_bed3 = f'enhancers_genes_tead4_tfap2c_dis_0based.bed3'
tfap2c_path = f'genes_tfap2c_dis_0based.bed'
tead4_path = f'genes_tead4_dis_0based.bed'
prefix = 'tt_enhancers'
enhancer_bed_file = f'{prefix}.bed'
cwm_scan_prefix = f'cwm_scan/{prefix}'
enhancer_contrib_file = f"contrib/{prefix}_contrib.h5"

%%script false --no-raise-error
!mkdir -p {cwm_scan_prefix}
!mkdir -p bed
!mkdir -p contrib
!mkdir -p figures/1_{prefix}_tt_distance

# Generate contributions based on enhancer coordinates

Generate contributions, then run CWM-scanning to map all pertinent motifs.

!cut -f 1-3 {enhancers_path} > {enhancers_bed3}

%%script false --no-raise-error
!bpnet contrib --regions {enhancers_bed3} --fasta-file {fasta_file} --method deeplift --overwrite  {model_dir} {enhancer_contrib_file}

%%script false --no-raise-error
# For each task, CWM-scan the enhancers. The `example_idx` will be the same across tasks
for task in ['tead4','tfap2c','cdx2','gata3','yap1']:
    print(task)
    cwm_scan(modisco_dir = f'{modisco_dir}/{task}', 
             output_file = f'{cwm_scan_prefix}/enhancer_cwm_scan_instances.{task}.tsv.gz', 
             contrib_file = enhancer_contrib_file)     

# Collect annotated genomic loci

Import the regions for downstream analysis. Because the enhancers are already a bed file, no conversion is needed.

#Import regions and the genomic coordinates.

enhancers_df = BedTool(enhancers_path).to_dataframe().sort_values(['name']).reset_index(drop = True)
tfap2c_df = BedTool(tfap2c_path).to_dataframe().sort_values(['name']).reset_index(drop = True)
tead4_df = BedTool(tead4_path).to_dataframe().sort_values(['name']).reset_index(drop = True)
enhancers_bed = BedTool.from_dataframe(enhancers_df)
tfap2c_bed = BedTool.from_dataframe(tfap2c_df)
tead4_bed = BedTool.from_dataframe(tead4_df)

#Ensure everything is in the same order and unique
assert(all(enhancers_df.name.values==tfap2c_df.name.values)==all(enhancers_df.name.values==tead4_df.name.values), 
       'Either motifs occur more than once in region OR the motifs and enhancers are not in the same order')
    

enhancers_df

tfap2c_df

Import enhancers, motifs, and collect coordinates for sequence mutation.

#Assign in-window coordinates to these motifs (watch enhancer id!)
tfap2c_df['window_enhancer_start'] = enhancers_df.start
tead4_df['window_enhancer_start'] = enhancers_df.start
tfap2c_df['window_enhancer_end'] = enhancers_df.end
tead4_df['window_enhancer_end'] = enhancers_df.end

tfap2c_df['window_motif_start'] = [m.start - m.window_enhancer_start for i,m in tfap2c_df.iterrows()]
tead4_df['window_motif_start'] = [m.start - m.window_enhancer_start for i,m in tead4_df.iterrows()]

tfap2c_df['window_motif_end'] = [m.end - m.window_enhancer_start for i,m in tfap2c_df.iterrows()]
tead4_df['window_motif_end'] = [m.end - m.window_enhancer_start for i,m in tead4_df.iterrows()]

tfap2c_df['window_motif_center'] = np.floor((tfap2c_df.window_motif_end - tfap2c_df.window_motif_start)/2) + tfap2c_df.window_motif_start
tead4_df['window_motif_center'] = np.floor((tead4_df.window_motif_end - tead4_df.window_motif_start)/2) + tead4_df.window_motif_start

# Mark where the motif boundaries are and store in enhancer_df
enhancers_df['tfap2c_start_genomic'] = tfap2c_df.start
enhancers_df['tfap2c_end_genomic'] = tfap2c_df.end
enhancers_df['tead4_start_genomic'] = tead4_df.start
enhancers_df['tead4_end_genomic'] = tead4_df.end

enhancers_df['tfap2c_center_genomic'] = np.floor((tfap2c_df.end - tfap2c_df.start)/2) + tfap2c_df.start
enhancers_df['tead4_center_genomic'] = np.floor((tead4_df.end - tead4_df.start)/2) + tead4_df.start

#Mark distance between the two motifs, with tfap2c as the anchor.
enhancers_df['tead4_minus_tfap2c_distance'] = enhancers_df['tfap2c_center_genomic'] - enhancers_df['tead4_center_genomic']

#enhancers_df[enhancers_df.name==119]

fasta_file

tfap2c_df.head(n=15)

tfap2c_df.head(n=15)

Confirm that the sequences found by BPNet match the sequences based off of the genomic coordinates of the .fasta file. To do this, we need to collect sequences and hyper-contributions across each enhancer in order to mutate them at the targeted sites.

# Get sequence and contribution 
c = ContribFile(enhancer_contrib_file)
seqs = c.get_seq()
coords = c.get_ranges()
hyp_contrib = c.get_hyp_contrib()

x =tead4_bed[1]
x
#[one_hot_decode(i) for i in FastaExtractor(fasta_file)([Interval.from_pybedtools(i) for i in x])]
tfap2c_df

Create a check that looks at every sequence to make sure it is correct.

#Extract BPNet-contained sequences
tfap2c_bpnet_seqs = [one_hot_decode(seqs[row.name][row.window_motif_start:row.window_motif_end]) for i,row in tfap2c_df.iterrows()]
tfap2c_fasta_seqs = [one_hot_decode(i) for i in FastaExtractor(fasta_file)([Interval.from_pybedtools(i) for i in tfap2c_bed])]
#assert tfap2c_bpnet_seqs==tfap2c_fasta_seqs, 'BPNet sequences and genomic sequences do not match. Correct this before proceeding.'

#tead4_bpnet_seqs = [one_hot_decode(seqs[row.name][row.window_motif_start:row.window_motif_end]) for i,row in tead4_df.iterrows()]
#tead4_fasta_seqs = [one_hot_decode(i) for i in FastaExtractor(fasta_file)([Interval.from_pybedtools(i) for i in tead4_bed])]
#assert tead4_bpnet_seqs==tead4_fasta_seqs, 'BPNet sequences and genomic sequences do not match. Correct this before proceeding.'

# Begin generating perturbations based on 2-mutation dynamic approach

TODO: Explain approach in normal words.

Define distances that we want to inject motifs across.

distances = range(-200, 200, 1)
mutation_count = 2 #how many bases we will mutate in a targeted sense

## Generate all sequences 

Across every enhancer, motif, and distance generate the WT, mutant, and mutant-distance injection. Combine into array while storing index information in a separate `pd.df`.

#Preallocate objects for writing.
wt_seqs = []
mut_seqs = []
mut_inj_seqs = []
wt_info_all_df = pd.DataFrame()
mut_inj_info_all_df = pd.DataFrame()

#test = enhancers_df.head(n=118)
for enhancer_id,enhancer_row in tqdm(enhancers_df.iterrows()):
    
    for anchor_motif in ['tfap2c', 'tead4']:
        #Define motif coordinates
        if anchor_motif=='tfap2c':
            anchor_motif_coords = tfap2c_df.loc[enhancer_id] 
            moved_motif = 'tead4'
            moved_motif_coords = tead4_df.loc[enhancer_id]
        else:
            anchor_motif_coords = tead4_df.loc[enhancer_id]
            moved_motif = 'tfap2c'
            moved_motif_coords = tfap2c_df.loc[enhancer_id]
        #Define input sequence across enhancer
        wt_seq = seqs[enhancer_id] 
        wt_seqs.append(wt_seq)
        
        #Find the minimum hyp contribution across all positions
        hyp_contrib_across_motif_arr = hyp_contrib[moved_motif][enhancer_id][moved_motif_coords.window_motif_start:moved_motif_coords.window_motif_end]
        min_contrib_across_motif_arr = np.min(hyp_contrib_across_motif_arr, axis = 1)
        min_indices_across_motif_arr = np.stack([np.where(pos==min_contrib_across_motif_arr[i])[0].flatten() for i,pos in enumerate(hyp_contrib_across_motif_arr)], axis = 0).flatten()
        seq_across_motif_arr = wt_seq[moved_motif_coords.window_motif_start:moved_motif_coords.window_motif_end]

        #Find the top contributing indices across the motif
        contrib_across_motif_arr = np.sum(hyp_contrib_across_motif_arr * seq_across_motif_arr, axis = 1)
        top_contributing_indices = (-contrib_across_motif_arr).argsort()[:mutation_count]

        #Reassign the worst contributing nt for each mutation position.
        mut_motif_seq = []
        for i in range(seq_across_motif_arr.shape[0]):
            if i in top_contributing_indices:
                row = np.array([0,0,0,0], dtype = 'float32')
                row[min_indices_across_motif_arr[i]] = 1
            else: 
                row = seq_across_motif_arr[i]
            mut_motif_seq.append(row)
        mut_motif_seq = np.stack(mut_motif_seq)

        #Reinject the motif sequence back into the entire window
        mut_seq = wt_seq.copy()
        mut_seq[moved_motif_coords.window_motif_start:moved_motif_coords.window_motif_end] = mut_motif_seq
        mut_seqs.append(mut_seq)
  
        #Record all perturbation information in a single pd.df (defined by enhancer_id)
        wt_info_df = pd.DataFrame([enhancer_id, anchor_motif, anchor_motif_coords.window_motif_start, anchor_motif_coords.window_motif_end, anchor_motif_coords.window_motif_center,
                                moved_motif, moved_motif_coords.window_motif_start, moved_motif_coords.window_motif_end, moved_motif_coords.window_motif_center]).transpose()
        wt_info_df.columns = ['enhancer_id',
                           'anchor_motif','anchor_motif_window_start_0based','anchor_motif_window_end_0based','anchor_motif_window_center_0based',
                           'moved_motif','moved_motif_window_orig_start_0based','moved_motif_window_orig_end_0based','moved_motif_window_orig_center_0based']
        wt_info_df['rank_positions'] = ','.join([str(x) for x in top_contributing_indices])
        wt_info_df['injected_motif_seq'] = one_hot_decode(seq_across_motif_arr)
        wt_info_df['mutated_motif_seq'] = one_hot_decode(mut_motif_seq)
        wt_info_df['anchor_motif_seq'] = one_hot_decode(wt_seq[anchor_motif_coords.window_motif_start:anchor_motif_coords.window_motif_end])
        wt_info_all_df = wt_info_all_df.append(wt_info_df)
        
        for d in distances:
            
            #Add check to make sure distance injection doesn't move off window
            check1 = (d + anchor_motif_coords.window_motif_start) <= 0
            check2 = (d + anchor_motif_coords.window_motif_end) >= (anchor_motif_coords.window_enhancer_end - anchor_motif_coords.window_enhancer_start - 1) 
            if not (check1 or check2):

                #Inject original motif into the designated distance 
                mut_inj_seq = mut_seq.copy()

                #Define distance boundaries based on anchor center
                injected_moved_motif_center = int(anchor_motif_coords.window_motif_center + d)
                upstream_to_center_moved_motif_distance = moved_motif_coords.window_motif_center - moved_motif_coords.window_motif_start
                downstream_to_center_moved_motif_distance = moved_motif_coords.window_motif_end - moved_motif_coords.window_motif_center 
                injected_moved_motif_boundaries = (int(injected_moved_motif_center - upstream_to_center_moved_motif_distance), 
                                                   int(injected_moved_motif_center + downstream_to_center_moved_motif_distance))
                
                #Add check to make sure distance doesn't inject over another motif.
                inj_motif_int = pd.Interval(injected_moved_motif_boundaries[0], injected_moved_motif_boundaries[1])
                anchor_motif_int = pd.Interval(anchor_motif_coords.window_motif_start, anchor_motif_coords.window_motif_end)
                if not inj_motif_int.overlaps(anchor_motif_int):
                    #Add check to make sure your mutation and injected sequence cover the same coordinates
                    test_mut = mut_inj_seq[injected_moved_motif_boundaries[0]:injected_moved_motif_boundaries[1]].shape == seq_across_motif_arr.shape
                    assert(test_mut, 'You are injecting a misproportioned sequence, check coordinates.')
                    mut_inj_seq[injected_moved_motif_boundaries[0]:injected_moved_motif_boundaries[1]] = seq_across_motif_arr
                    mut_inj_seqs.append(mut_inj_seq)

                    #Record all perturbation information in a single pd.df (distance_id)
                    mut_inj_info_df = wt_info_df
                    mut_inj_info_df['distance_between_moved_center_and_anchor_center'] = d
                    mut_inj_info_df['distance_id'] = str(mut_inj_info_df.enhancer_id.values[0]) + '_' + str(d)
                    mut_inj_info_all_df = mut_inj_info_all_df.append(mut_inj_info_df)
                ##below commented out is the alternative approach to get rank, high contributing nucleotides and mutated nucleotides:
                # df = pd.DataFrame([top_contributing_indices, list(range(mutation_count)), 
                #              list(one_hot_decode(seq_across_motif_arr[top_contributing_indices])),
                #              list(one_hot_decode(mut_motif_seq[top_contributing_indices]))]).transpose()
                # df.columns = ['mutation_index_across_motif', 'mutation_rank', 'wt_nt','mut_nt']
                # df.pivot(columns='mutation_rank', values=['mutation_index_across_motif', 'wt_nt','mut_nt'])
wt_seqs = np.array(wt_seqs)
mut_seqs = np.array(mut_seqs)
mut_inj_seqs = np.array(mut_inj_seqs)

#Add checks to make sure correct dimensions match enhancer ids
assert wt_seqs.shape[0]==mut_seqs.shape[0]==wt_info_all_df.shape[0], 'WT and mut arrays dont match their index'
assert mut_inj_seqs.shape[0]==mut_inj_info_all_df.shape[0], 'WT and mut arrays dont match their index'


mut_inj_info_all_df.head

Export full seqs to csv.gz and make sure full seqs match.

wt_test_df = pd.DataFrame([list(one_hot_decode(wt_seqs[i])) for i in range(wt_seqs.shape[0])]).transpose()
mut_test_df = pd.DataFrame([list(one_hot_decode(mut_seqs[i])) for i in range(mut_seqs.shape[0])]).transpose()
mut_inj_test_df = pd.DataFrame([list(one_hot_decode(mut_inj_seqs[i])) for i in range(mut_inj_seqs.shape[0])]).transpose()
wt_test_df.to_csv('tmp/wt_seqs.tsv', sep = '\t')
mut_test_df.to_csv('tmp/mut_seqs.tsv', sep = '\t')
mut_inj_test_df.to_csv('tmp/mut_inj_seqs.tsv', sep = '\t')

# Predict each sequence set

Predict the WT, mutated, and mutated-injected sequences that were generated above. First, import the model of interest to generate predictions.

model = BPNetSeqModel.from_mdir(model_dir)

Get predictions from each set of sequences.

#Get predictions from sequences
wt_preds = model.predict(wt_seqs)
mut_preds = model.predict(mut_seqs)
mut_inj_preds = model.predict(mut_inj_seqs)

Measure the whole-window counts summary of these perturbations, and link them into a single data frame. Add motif-range measurements in addition to whole-window measurements. Make this dynamic based on the injection site of distance.

First, define the window range for the motif-range measurements.

motif_range_width = 50

Next, collect measurements.

#Measure WT and mutant task predictions across the whole window
wt_preds_df = pd.DataFrame.from_dict({k: np.sum(v, axis = (1,2)) for k,v in wt_preds.items()})
wt_preds_df.columns = [i + '_wt_whole-range' for i in wt_preds_df.columns]
mut_preds_df = pd.DataFrame.from_dict({k:np.sum(v, axis = (1,2)) for k,v in mut_preds.items()})
mut_preds_df.columns = [i + '_mut_whole-range' for i in mut_preds_df.columns]
wt_info_w_preds_df = pd.concat([wt_info_all_df.reset_index(), wt_preds_df.reset_index(), mut_preds_df.reset_index()], axis=1)

wt_info_w_preds_df.shape[0]

#Measure mut_inj predictions across the whole window for every distance
mut_inj_preds_df = pd.DataFrame.from_dict({k: np.sum(v, axis = (1,2)) for k,v in mut_inj_preds.items()})
mut_inj_preds_df.columns = [i + '_mut-inj_whole-range' for i in mut_inj_preds_df.columns]
mut_inj_info_w_preds_df = pd.concat([mut_inj_info_all_df.reset_index(), mut_inj_preds_df.reset_index()], axis=1)
mut_inj_info_w_preds_df['genomic_moved_minus_anchor_motif_distance'] = mut_inj_info_w_preds_df.moved_motif_window_orig_center_0based - mut_inj_info_w_preds_df.anchor_motif_window_center_0based
whole_window_summary_df = mut_inj_info_w_preds_df.merge(wt_info_w_preds_df[['enhancer_id', 'anchor_motif', 'moved_motif'] + list(wt_preds_df.columns) +  list(mut_preds_df.columns)], 
                              on=['enhancer_id', 'anchor_motif', 'moved_motif'], how = 'left')

mut_inj_info_w_preds_df.shape[0]

For each set of enhancers, anchor/move motifs, and distances, extract the motif-window measurements.

motif_window_predictions = pd.DataFrame()
for i,row in tqdm(whole_window_summary_df.iterrows()):
    injected_center = row.anchor_motif_window_center_0based - row.distance_between_moved_center_and_anchor_center
    injected_coords = (injected_center - (row.moved_motif_window_orig_center_0based - row.moved_motif_window_orig_start_0based),
                       injected_center - (row.moved_motif_window_orig_end_0based - row.moved_motif_window_orig_center_0based))
    injected_motif_window = (int(injected_center - np.floor(motif_range_width/2)), 
                             int(injected_center + np.floor(motif_range_width/2)))
    
    #Define enhancer_id and anchor_motif to index WT_preds and MUT_preds
    enhancer_id = row.enhancer_id
    anchor_motif = row.anchor_motif
    wt_and_mut_index = int(wt_info_w_preds_df[(wt_info_w_preds_df.enhancer_id==enhancer_id) & (wt_info_w_preds_df.anchor_motif==anchor_motif)].index.values[0])
    print(wt_and_mut_index)                   
    #Measure motif range across 3 types for injected window
    i_wt_preds_df = pd.DataFrame({k: np.sum(v[wt_and_mut_index, injected_motif_window[0]:injected_motif_window[1], :], axis = (0,1)) 
                                          for k,v in wt_preds.items()}, index=[0])
    i_wt_preds_df.columns = [i + '_wt_motif-range_inj-window' for i in i_wt_preds_df.columns]
    i_mut_preds_df = pd.DataFrame({k:np.sum(v[wt_and_mut_index, injected_motif_window[0]:injected_motif_window[1], :], axis = (0,1)) 
                                           for k,v in mut_preds.items()}, index=[0])
    i_mut_preds_df.columns = [i + '_mut_motif-range_inj-window' for i in i_mut_preds_df.columns]    
    i_mut_inj_preds_df = pd.DataFrame({k: np.sum(v[i, injected_motif_window[0]:injected_motif_window[1], :], axis = (0,1)) 
                                               for k,v in mut_inj_preds.items()}, index=[0])
    i_mut_inj_preds_df.columns = [i + '_mut-inj_motif-range_inj-window' for i in i_mut_inj_preds_df.columns]
                      
    #Measure motif range across 3 types for anchor window
    anchor_motif_window = (int(row.anchor_motif_window_center_0based - np.floor(motif_range_width/2)), 
                           int(row.anchor_motif_window_center_0based + np.floor(motif_range_width/2)))                            
    a_wt_preds_df = pd.DataFrame({k: np.sum(v[wt_and_mut_index, anchor_motif_window[0]:anchor_motif_window[1], :], axis = (0,1)) 
                                          for k,v in wt_preds.items()}, index=[0])
    a_wt_preds_df.columns = [i + '_wt_motif-range_anchor-window' for i in a_wt_preds_df.columns]
    a_mut_preds_df = pd.DataFrame({k:np.sum(v[wt_and_mut_index, anchor_motif_window[0]:anchor_motif_window[1], :], axis = (0,1)) 
                                           for k,v in mut_preds.items()}, index=[0])
    a_mut_preds_df.columns = [i + '_mut_motif-range_anchor-window' for i in a_mut_preds_df.columns]   
    a_mut_inj_preds_df = pd.DataFrame({k: np.sum(v[i, anchor_motif_window[0]:anchor_motif_window[1], :], axis = (0,1)) 
                                               for k,v in mut_inj_preds.items()}, index=[0])
    a_mut_inj_preds_df.columns = [i + '_mut-inj_motif-range_anchor-window' for i in a_mut_inj_preds_df.columns]
    
    #Measure motif range across 3 types for mutation window
    moved_orig_motif_window = (int(row.moved_motif_window_orig_center_0based - np.floor(motif_range_width/2)), 
                               int(row.moved_motif_window_orig_center_0based + np.floor(motif_range_width/2)))                            
    m_wt_preds_df = pd.DataFrame({k: np.sum(v[wt_and_mut_index, moved_orig_motif_window[0]:moved_orig_motif_window[1], :], axis = (0,1)) 
                                          for k,v in wt_preds.items()}, index=[0])
    m_wt_preds_df.columns = [i + '_wt_motif-range_mut-window' for i in m_wt_preds_df.columns]
    m_mut_preds_df = pd.DataFrame({k:np.sum(v[wt_and_mut_index, moved_orig_motif_window[0]:moved_orig_motif_window[1], :], axis = (0,1)) 
                                           for k,v in mut_preds.items()}, index=[0])
    m_mut_preds_df.columns = [i + '_mut_motif-range_mut-window' for i in m_mut_preds_df.columns]    
    m_mut_inj_preds_df = pd.DataFrame({k: np.sum(v[i, moved_orig_motif_window[0]:moved_orig_motif_window[1], :], axis = (0,1)) 
                                               for k,v in mut_inj_preds.items()}, index=[0])
    m_mut_inj_preds_df.columns = [i + '_mut-inj_motif-range_mut-window' for i in m_mut_inj_preds_df.columns]
                               
    motif_window_prediction = pd.concat([a_wt_preds_df, a_mut_preds_df, a_mut_inj_preds_df,
                                         m_wt_preds_df, m_mut_preds_df, m_mut_inj_preds_df,
                                         i_wt_preds_df, i_mut_preds_df, i_mut_inj_preds_df],axis = 1)  

    motif_window_predictions = motif_window_predictions.append(motif_window_prediction)
                               
assert motif_window_predictions.shape[0]==whole_window_summary_df.shape[0]
final_pred_summary_df = pd.concat([whole_window_summary_df,motif_window_predictions.reset_index()], axis = 1)                              

final_pred_summary_df.head(n=5)

## save each enhancer_id as seperate tsv
for id in final_pred_summary_df['enhancer_id'].unique():
    subset_df = final_pred_summary_df[final_pred_summary_df['enhancer_id'] == id]
    filename = f'two_mutation/tsv/individual_pred_summary_by_enhancer_id/' + str(id) + '_pred_summary_df' + '.tsv.gz'
    subset_df.to_csv(filename, index=False, sep='\t')  

#save `final_pred_summary_df` as .tsv.gz
final_pred_summary_df.to_csv(f'two_mutation/tsv/combined/{prefix}_final_pred_summary_df.tsv.gz', sep = '\t', index = False)
final_pred_summary_df.shape

%%script false --no-raise-error
#read csv
final_pred_summary_df= pd.read_csv(f'two_mutation/tsv/combined/{prefix}_final_pred_summary_df.tsv.gz', sep = '\t')
final_pred_summary_df.shape

Measure entire profile signals across desired distances of mut/inj, mut, and wt predictions. Store in a tidy `pd.df` and export to a `tsv.gz` for plotting using `plotnine` or `ggplot` in R. First, specify distances that you want to preserve based on injection distance from the anchor motif.

distances_to_store_profile = list(range(0, 200, 15)) + [i * -1 for i in list(range(0, 200, 15))]
distances_to_store_profile

#distances_to_store_profile = list(range(0, 200, 10)) + [i * -1 for i in list(range(0, 200, 10))]
#distances_to_store_profile

Next, extract and tidy profiles, eventually saving them as a `.tsv.gz`.

all_profiles_df = pd.DataFrame()

for task in model.tasks:
    ## Generate WT profiles
    #Allocate all positive reads as 2D matrix and then a data frame
    wt_pos_profile_df = pd.DataFrame(data=wt_preds[task][:,:,0], columns=list(range(0, model.input_seqlen())))
    wt_pos_profile_df['enhancer_id'] = wt_info_w_preds_df.enhancer_id
    wt_pos_profile_df['anchor_motif'] = wt_info_w_preds_df.anchor_motif
    wt_pos_profile_df['moved_motif'] = wt_info_w_preds_df.moved_motif
    wt_pos_profile_df = wt_pos_profile_df.melt(id_vars = ['enhancer_id', 'anchor_motif', 'moved_motif'], 
                                               var_name = 'position', value_name = 'signal')
    wt_pos_profile_df['type'] = 'wt'
    wt_pos_profile_df['strand'] = 'pos'
    wt_pos_profile_df['task'] = task

    #Allocate all negative reads as 2D matrix and then a data frame
    wt_neg_profile_df = pd.DataFrame(data=wt_preds[task][:,:,1], columns=list(range(0, model.input_seqlen())))
    wt_neg_profile_df['enhancer_id'] = wt_info_w_preds_df.enhancer_id
    wt_neg_profile_df['anchor_motif'] = wt_info_w_preds_df.anchor_motif
    wt_neg_profile_df['moved_motif'] = wt_info_w_preds_df.moved_motif
    wt_neg_profile_df = wt_neg_profile_df.melt(id_vars = ['enhancer_id', 'anchor_motif', 'moved_motif'], 
                                               var_name = 'position', value_name = 'signal')
    wt_neg_profile_df['type'] = 'wt'
    wt_neg_profile_df['strand'] = 'neg'
    wt_neg_profile_df['signal'] = -1 * wt_neg_profile_df['signal']
    wt_neg_profile_df['task'] = task

    ## Generate mut profiles
    #Allocate all positive reads as 2D matrix and then a data frame
    mut_pos_profile_df = pd.DataFrame(data=mut_preds[task][:,:,0], columns=list(range(0, model.input_seqlen())))
    mut_pos_profile_df['enhancer_id'] = wt_info_w_preds_df.enhancer_id
    mut_pos_profile_df['anchor_motif'] = wt_info_w_preds_df.anchor_motif
    mut_pos_profile_df['moved_motif'] = wt_info_w_preds_df.moved_motif
    mut_pos_profile_df = mut_pos_profile_df.melt(id_vars = ['enhancer_id', 'anchor_motif', 'moved_motif'], 
                                                 var_name = 'position', value_name = 'signal')
    mut_pos_profile_df['type'] = 'mut'
    mut_pos_profile_df['strand'] = 'pos'
    mut_pos_profile_df['task'] = task

    #Allocate all negative reads as 2D matrix and then a data frame
    mut_neg_profile_df = pd.DataFrame(data=mut_preds[task][:,:,1], columns=list(range(0, model.input_seqlen())))
    mut_neg_profile_df['enhancer_id'] = wt_info_w_preds_df.enhancer_id
    mut_neg_profile_df['anchor_motif'] = wt_info_w_preds_df.anchor_motif
    mut_neg_profile_df['moved_motif'] = wt_info_w_preds_df.moved_motif
    mut_neg_profile_df = mut_neg_profile_df.melt(id_vars = ['enhancer_id', 'anchor_motif', 'moved_motif'], 
                                                 var_name = 'position', value_name = 'signal')
    mut_neg_profile_df['type'] = 'mut'
    mut_neg_profile_df['strand'] = 'neg'
    mut_neg_profile_df['signal'] = -1 * mut_neg_profile_df['signal']
    mut_neg_profile_df['task'] = task

    ## Generate mut and injectedprofiles
    #Allocate all positive reads as 2D matrix and then a data frame
    mut_inj_pos_profile_df = pd.DataFrame(data=mut_inj_preds[task][:,:,0], columns=list(range(0, model.input_seqlen())))
    mut_inj_pos_profile_df['enhancer_id'] = mut_inj_info_w_preds_df.enhancer_id
    mut_inj_pos_profile_df['anchor_motif'] = mut_inj_info_w_preds_df.anchor_motif
    mut_inj_pos_profile_df['moved_motif'] = mut_inj_info_w_preds_df.moved_motif
    mut_inj_pos_profile_df['type'] = ['mut_inj/' +  str(i) for i in mut_inj_info_w_preds_df.distance_between_moved_center_and_anchor_center.values]
    mut_inj_pos_profile_df = mut_inj_pos_profile_df.melt(id_vars = ['enhancer_id', 'anchor_motif', 'moved_motif', 'type'], 
                                                         var_name = 'position', value_name = 'signal')
    mut_inj_pos_profile_df['strand'] = 'pos'
    mut_inj_pos_profile_df['task'] = task
    
    #Label distances desired to store
    distances_to_store_profile_label = ['mut_inj/' +  str(i) for i in distances_to_store_profile]

    #Filter out distances not desired to store        
    mut_inj_pos_profile_df = mut_inj_pos_profile_df[mut_inj_pos_profile_df['type'].isin(list(distances_to_store_profile_label))]
    
    #Allocate all negative reads as 2D matrix and then a data frame
    mut_inj_neg_profile_df = pd.DataFrame(data=mut_inj_preds[task][:,:,1], columns=list(range(0, model.input_seqlen())))
    mut_inj_neg_profile_df['enhancer_id'] = mut_inj_info_w_preds_df.enhancer_id
    mut_inj_neg_profile_df['anchor_motif'] = mut_inj_info_w_preds_df.anchor_motif
    mut_inj_neg_profile_df['moved_motif'] = mut_inj_info_w_preds_df.moved_motif
    mut_inj_neg_profile_df['type'] = ['mut_inj/' +  str(i) for i in mut_inj_info_w_preds_df.distance_between_moved_center_and_anchor_center.values]
    mut_inj_neg_profile_df = mut_inj_neg_profile_df.melt(id_vars = ['enhancer_id', 'anchor_motif', 'moved_motif', 'type'], 
                                                         var_name = 'position', value_name = 'signal')
    mut_inj_neg_profile_df['strand'] = 'neg'
    mut_inj_neg_profile_df['signal'] = -1 * mut_inj_neg_profile_df['signal']
    mut_inj_neg_profile_df['task'] = task
    mut_inj_neg_profile_df = mut_inj_neg_profile_df[mut_inj_neg_profile_df['type'].isin(list(distances_to_store_profile_label))]
    
    all_profile_df = pd.concat([wt_pos_profile_df, wt_neg_profile_df, mut_pos_profile_df, mut_neg_profile_df, 
                                mut_inj_pos_profile_df, mut_inj_neg_profile_df], axis = 0).reset_index()
    all_profiles_df = all_profiles_df.append(all_profile_df)


#Save as tsv.gz
all_profiles_df.to_csv(f'two_mutation/tsv/combined/{prefix}_all_profiles_multiples_15_df.tsv.gz', sep = '\t', index = False)
all_profiles_df.shape

pdf= pd.read_csv(f'two_mutation/tsv/combined/{prefix}_all_profiles_multiples_15_df.tsv.gz',sep = '\t')

pdf

## save each enhancer_id as seperate tsv
for id in pdf['enhancer_id'].unique():
    subset_df = pdf[pdf['enhancer_id'] == id]
    filename = f'two_mutation/tsv/individual_profile_by_enhancer_id_dis_15_multiples/' + str(id) + '_profiles_df' + '.tsv.gz'
    subset_df.to_csv(filename, index=False, sep='\t')
